fixed the cells, display of jobs and added and extensions for the jobs table view
