+++ letsApply/Controllers/TabBarViewController/JobListingsViewController/JobListViewController/JobListingsViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 
 class JobListingsViewController: UIViewController {
     
-    private let viewModel = JobListingsViewModel()
+    lazy var viewModel = JobListingsViewModel()
     
     // UI Elements
     lazy var tableView: UITableView = {
@@ -65,28 +65,6 @@ class JobListingsViewController: UIViewController {
         let filtersVC = JobFiltersViewController()
         filtersVC.delegate = self
         navigationController?.pushViewController(filtersVC, animated: true)
-    }
-}
-
-extension JobListingsViewController: UITableViewDataSource, UITableViewDelegate {
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return viewModel.jobs.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: JobTableViewCell.identifier, for: indexPath) as? JobTableViewCell else {
-            return UITableViewCell()
-        }
-        let job = viewModel.jobs[indexPath.row]
-        cell.configure(with: job)
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let job = viewModel.jobs[indexPath.row]
-        let jobDetailsVC = JobDetailsViewController(job: job)
-        navigationController?.pushViewController(jobDetailsVC, animated: true)
     }
 }
