+++ letsApply/Extensions/UITableViewCell+Extension.swift
@@ -1,0 +1,15 @@
+//
+//  UITableViewCell+Extension.swift
+//  letsApply
+//
+//  Created by Gugulethu Mhlanga on 2024/11/30.
+//
+
+import Foundation
+import UIKit
+
+extension UITableViewCell {
+    static var identifier: String {
+        return String(describing: self)
+    }
+}
